# api/openapi.yaml
openapi: 3.0.3
info:
  title: Weave Ontology API
  version: 1.0.0
  description: |
    Weave is a single-machine Ontology Layer engine inspired by Palantir Foundry OSv2.
    This document describes its REST API: ontology metadata, object retrieval and
    search, ObjectSet composition, aggregation, link traversal, action execution,
    and authentication.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (port configurable via WEAVE_PORT).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256-signed JWT issued via /api/auth/login. Send as
        `Authorization: Bearer <token>`.
    apiKey:
      type: apiKey
      in: header
      name: X-Weave-Api-Key
      description: Reserved for future API-key-based service-to-service auth.
  responses:
    BadRequestError:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFoundError:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  parameters:
    OntologyApiName:
      in: path
      name: ontologyApiName
      required: true
      description: API name (or RID) of the ontology.
      schema:
        type: string
...
    ObjectType:
      in: path
      name: objectType
      required: true
      description: API name of the object type.
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      description: Palantir-style API error envelope.
      required: [errorCode, errorName, errorInstanceId, parameters]
      properties:
        errorCode:
          type: string
          example: NOT_FOUND
        errorName:
          type: string
          example: OntologyNotFound
        errorInstanceId:
          type: string
          format: uuid
        parameters:
          type: object
          additionalProperties:
            type: string
...
    ObjectSetDefinition:
      type: object
      description: |
        Composable ObjectSet definition. Supported types include base, filter,
        union, intersect, subtract, and searchAround. See pkg/oss/objectset.
      additionalProperties: true
      properties:
        type:
          type: string
    LoadObjectSetRequest:
      type: object
      required: [objectSet]
      properties:
        objectSet:
          $ref: "#/components/schemas/ObjectSetDefinition"
        select:
          type: array
          items:
            type: string
        orderBy:
          type: object
          properties:
            fields:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  direction:
                    type: string
                    enum: [asc, desc]
        pageSize:
          type: integer
        pageToken:
          type: string
        snapshot:
          type: boolean
    LoadObjectSetResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WireObject"
        nextPageToken:
          type: string
        totalCount:
          type: string
        totalCountAccuracy:
          type: string
          enum: [EXACT, APPROXIMATE]
    AggregationSpec:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [count, min, max, sum, avg, approximateDistinct, standardDeviation, variance, approximatePercentile]
        field:
          type: string
        name:
          type: string
        percentile:
          type: number
    GroupBySpec:
      type: object
      required: [type, field]
      properties:
        type:
          type: string
          enum: [exact, fixedWidth, range, ranges, duration, topValues]
        field:
          type: string
        maxGroupCount:
          type: integer
        fixedWidth:
          type: number
        ranges:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              startValue:
                type: number
              endValue:
                type: number
        duration:
          type: string
        value:
          type: object
          properties:
            unit:
              type: string
              enum: [SECONDS, MINUTES, HOURS, DAYS, WEEKS, MONTHS, YEARS]
            value:
              type: number
    AggregationRequest:
      type: object
      required: [aggregation]
      properties:
        objectType:
          type: string
        aggregation:
          type: array
          items:
            $ref: "#/components/schemas/AggregationSpec"
        groupBy:
          type: array
          items:
            $ref: "#/components/schemas/GroupBySpec"
    AggregationRow:
      type: object
      properties:
        group:
          type: object
          additionalProperties: true
        metrics:
          type: object
          additionalProperties: true
    AggregationResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AggregationRow"
        accuracy:
          type: string
          enum: [ACCURATE, APPROXIMATE]
        computeUsage:
          type: number
    AggregateObjectSetRequest:
      type: object
      required: [objectSet, aggregation]
      properties:
        objectSet:
          $ref: "#/components/schemas/ObjectSetDefinition"
        aggregation:
          type: array
          items:
            $ref: "#/components/schemas/AggregationSpec"
        groupBy:
          type: array
          items:
            $ref: "#/components/schemas/GroupBySpec"
    CreateTemporaryRequest:
      type: object
      required: [objectSet]
      properties:
        objectSet:
          $ref: "#/components/schemas/ObjectSetDefinition"
    CreateTemporaryResponse:
      type: object
      properties:
        objectSetRid:
          type: string
...
    ApplyActionRequest:
      type: object
      required: [actionType]
      properties:
        actionType:
          type: string
        parameters:
          type: object
          additionalProperties: true
    ApplyActionResponse:
      type: object
      properties:
        actionRid:
          type: string
        edits:
          type: array
          items:
            $ref: "#/components/schemas/Edit"
        batchId:
          type: string
        offset:
          type: integer
          format: int64
...
    ApplyBatchRequest:
      type: object
      required: [actions]
      properties:
        actions:
          type: array
          items:
            $ref: "#/components/schemas/ApplyActionRequest"
        mode:
          type: string
          enum: [atomic, bestEffort]
          default: atomic
    ApplyBatchResponse:
      type: object
      properties:
        mode:
          type: string
        batchId:
          type: string
        offset:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: "#/components/schemas/ApplyActionResponse"
        appliedEdits:
          type: array
          items:
            $ref: "#/components/schemas/Edit"
        failures:
          type: array
          items:
            $ref: "#/components/schemas/BatchFailure"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginUser:
      type: object
      required: [id, email, name, roles, ontologyRoles]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        ontologyRoles:
          type: object
          additionalProperties:
            type: string
    LoginResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
        user:
          $ref: "#/components/schemas/LoginUser"
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      $ref: "#/components/schemas/LoginResponse"
    User:
      type: object
      required: [id, roles, ontologyRoles, permissions]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        ontologyRoles:
          type: object
          additionalProperties:
            type: string
        permissions:
          type: array
          items:
            type: string
...
security:
  - bearerAuth: []
paths:
...
  /mcp:
    post:
      operationId: mcpJSONRPC
      summary: Model Context Protocol JSON-RPC 2.0 endpoint
      description: |
        Single JSON-RPC 2.0 endpoint that exposes Weave's ontology to AI
        agents (Claude Desktop, Cursor, GPT, etc.) via the Model Context
        Protocol. Supported methods:

        - `initialize` — handshake; returns protocol version, capabilities, and serverInfo
        - `tools/list` — list registered Weave tools and their input schemas
        - `tools/call` — invoke a tool by name with arguments
        - `prompts/list` — returns an empty list (Weave does not yet expose prompts)
        - `resources/list` — returns an empty list
        - `ping` — liveness check

        The seven Weave tools are: `weave_list_ontologies`,
        `weave_list_object_types`, `weave_get_object`, `weave_list_objects`,
        `weave_search_objects`, `weave_list_action_types`, and
        `weave_apply_action`. See docs/mcp.md for the full catalogue.

        Per JSON-RPC 2.0, all protocol errors (parse error, invalid request,
        method not found, invalid params, internal error, application
        error) are returned in the response envelope with HTTP 200. Non-200
        status codes are reserved for transport-level failures.
      tags: [MCP]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: |
                JSON-RPC 2.0 request envelope. The `id` field is required for
                requests that expect a response and omitted for notifications.
              required: [jsonrpc, method]
              properties:
                jsonrpc:
                  type: string
                  enum: ["2.0"]
                id:
                  description: Request id (string, number, or null). Omitted for notifications.
                method:
                  type: string
                  example: tools/call
                params:
                  type: object
      responses:
        "200":
          description: |
            JSON-RPC 2.0 response envelope. Contains either a `result` or an
            `error` object — never both. Application errors travel here too.
          content:
            application/json:
              schema:
                type: object
                required: [jsonrpc]
                properties:
                  jsonrpc:
                    type: string
                    enum: ["2.0"]
                  id:
                    description: Echo of the request id (or null).
                  result:
                    type: object
                    description: Method-specific result. Present on success.
                  error:
                    type: object
                    description: JSON-RPC error object. Present on failure.
                    properties:
                      code:
                        type: integer
                      message:
                        type: string
                      data: {}
        "204":
          description: Notification accepted (request had no `id` field).
        "405":
          description: HTTP method other than POST.
  /api/auth/login:
    post:
      operationId: login
      summary: Issue access and refresh tokens
      description: |
        Validates credentials, mints an RS256 JWT access token and a rotating
        refresh token. Returns the user's roles, scoped ontology roles, and
        token TTL.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      operationId: refresh
      summary: Rotate refresh token and mint a new access token
      description: |
        Rotates the supplied refresh token (from body or
        `weave_refresh` cookie) and returns a new access + refresh pair with
        freshly resolved roles.
      tags: [Auth]
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Refresh successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/auth/logout:
    post:
      operationId: logout
      summary: Revoke a refresh token
      description: |
        Revokes the supplied refresh token (or the cookie value), clears the
        refresh cookie, and always returns 204. Idempotent.
      tags: [Auth]
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Logout completed.
  /api/v2/me:
    get:
      operationId: getMe
      summary: Return the authenticated user
      description: Returns the resolved roles, scoped roles, and permission set.
      tags: [Auth]
      responses:
        "200":
          description: Authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/v2/ontologies/{ontologyApiName}/objects/{objectType}/aggregate:
    post:
      operationId: aggregateObjects
      summary: Aggregate objects via Bleve facets
      tags: [Objects]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
        - $ref: "#/components/parameters/ObjectType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AggregationRequest"
      responses:
        "200":
          description: Aggregation result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregationResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v2/ontologies/{ontologyApiName}/objectSets/loadObjects:
    post:
      operationId: loadObjectSet
      summary: Materialize an ObjectSet definition into a page of objects
      tags: [ObjectSets]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoadObjectSetRequest"
      responses:
        "200":
          description: ObjectSet load response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoadObjectSetResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v2/ontologies/{ontologyApiName}/objectSets/aggregate:
    post:
      operationId: aggregateObjectSet
      summary: Aggregate objects in an ObjectSet
      tags: [ObjectSets]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AggregateObjectSetRequest"
      responses:
        "200":
          description: Aggregation result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregationResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v2/ontologies/{ontologyApiName}/objectSets/createTemporary:
    post:
      operationId: createTemporaryObjectSet
      summary: Cache an ObjectSet definition for later reuse
      tags: [ObjectSets]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemporaryRequest"
      responses:
        "200":
          description: Temporary ObjectSet handle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTemporaryResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/v2/ontologies/{ontologyApiName}/actions/apply:
    post:
      operationId: applyAction
      summary: Apply a single action
      tags: [Actions]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyActionRequest"
      responses:
        "200":
          description: Action result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyActionResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v2/ontologies/{ontologyApiName}/actions/applyBatch:
    post:
      operationId: applyActionBatch
      summary: Apply a batch of actions atomically or best-effort
      tags: [Actions]
      parameters:
        - $ref: "#/components/parameters/OntologyApiName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyBatchRequest"
      responses:
        "200":
          description: Batch result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyBatchResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...